refactor: utils index
